requirements.txt
# Core dependencies
streamlit==1.31.1
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
langchain==0.1.11
openai==1.12.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
numpy==1.26.4
matplotlib==3.8.3
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1  # For progress bars

# Development dependencies
black==23.12.1
flake8==6.1.0
pytest==7.4.3

# For OpenRouter integration
openrouter==1.0.0
